backend/requirements.txt
flask
torch
torchvision
opencv-python
Pillow
numpy
